requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.2
asgiref==3.4.1
atpublic==2.3
attrs==21.2.0
cffi==1.14.6
Django==3.2.6
django-phonenumber-field==5.2.0
django-summernote==0.8.11.6
Django-Verify-Email==1.0.7
phonenumbers==8.12.31
Pillow==8.3.1
pycares==4.0.0
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
verify-email==2.4.3
